requirements.txt
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
leaflet==0.0.3
MarkupSafe==1.1.1
pandas==1.1.3
plotly==4.11.0
psycopg2==2.8.6
pymongo==3.11.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
retrying==1.3.3
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1
python-dotenv==0.15.0